A few updates to support yubikeys and az win
Some of us are storing our ssh keys on our yubikeys and no longer have
an id_rsa.pub file on disk.  I've updated the script to take in args
and allow for identifying that a yubikey is present.  It will try to
get an ssh key as best it can.

az cli returns CR at the end of the line if the windows version is used
in wsl, I fix that up if that is the case.

Default resource group is AnalyticalPlatform

Have a usage function to explain the options in the script

filter down the results from kubectl get nodes -w to just the nodes
related to the scale set name.